Merge staging to main finishing sprint

+++ crm-project/src/main/java/CRM/controller/facades/NotificationFacade.java
@@ -6,9 +6,12 @@ import CRM.entity.requests.ObjectsIdsRequest;
 import CRM.entity.response.Response;
 import CRM.service.NotificationService;
 import CRM.utils.Validations;
+import CRM.utils.enums.ExceptionMessage;
 import CRM.utils.enums.Regex;
 import CRM.utils.enums.SuccessMessage;
 import com.google.api.client.http.HttpStatusCodes;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
@@ -19,6 +22,8 @@ import java.util.NoSuchElementException;
 
 @Component
 public class NotificationFacade {
+
+    private static Logger logger = LogManager.getLogger(NotificationFacade.class.getName());
 
     @Autowired
     private NotificationService notificationService;
@@ -33,29 +38,14 @@ public class NotificationFacade {
      * @throws NoSuchElementException   if the specified board does not exist
      * @throws NullPointerException     if the objects IDs request object is null
      */
-    public Response getAllinBoard(ObjectsIdsRequest objectsIdsRequest) {
-        try {
-            Validations.validateIDs(objectsIdsRequest.getUserId(), objectsIdsRequest.getBoardId());
-            List<Notification> notificationList = notificationService.getAllinBoard(objectsIdsRequest);
-            return Response.builder()
-                    .message(SuccessMessage.FOUND.toString())
-                    .status(HttpStatus.FOUND)
-                    .statusCode(HttpStatusCodes.STATUS_CODE_OK)
-                    .data(NotificationDTO.createNotificationsDTOList(notificationList))
-                    .build();
-        } catch (AccountNotFoundException | IllegalArgumentException | NoSuchElementException e) {
-            return Response.builder()
-                    .statusCode(HttpStatusCodes.STATUS_CODE_BAD_REQUEST)
-                    .status(HttpStatus.BAD_REQUEST)
-                    .message(e.getMessage())
-                    .build();
-        } catch (NullPointerException e) {
-            return Response.builder()
-                    .statusCode(HttpStatusCodes.STATUS_CODE_SERVER_ERROR)
-                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .message(e.getMessage())
-                    .build();
-        }
+    public Response<List<NotificationDTO>> getAllinBoard(ObjectsIdsRequest objectsIdsRequest) throws AccountNotFoundException {
+        Validations.validateIDs(objectsIdsRequest.getUserId(), objectsIdsRequest.getBoardId());
+        return Response.<List<NotificationDTO>>builder()
+                .message(SuccessMessage.FOUND.toString())
+                .status(HttpStatus.FOUND)
+                .statusCode(HttpStatusCodes.STATUS_CODE_OK)
+                .data(NotificationDTO.createNotificationsDTOList(notificationService.getAllinBoard(objectsIdsRequest)))
+                .build();
     }
 
     /**
@@ -67,33 +57,21 @@ public class NotificationFacade {
      * @throws NoSuchElementException   if any of the specified notifications do not exist
      * @throws NullPointerException     if the notifications IDs list is null
      */
-    public Response delete(List<Long> notificationsIds) {
-        try {
-            notificationsIds.forEach(id -> {
-                try {
-                    Validations.validate(id, Regex.ID.getRegex());
-                } catch (IllegalArgumentException | NullPointerException e) {
-                }
-            });
-            List<Notification> remainingNotifications = notificationService.delete(notificationsIds);
-            return Response.builder()
-                    .message(SuccessMessage.DELETED.toString())
-                    .status(HttpStatus.OK)
-                    .statusCode(204)
-                    .data(NotificationDTO.createNotificationsDTOList(remainingNotifications))
-                    .build();
-        } catch (IllegalArgumentException | NoSuchElementException e) {
-            return Response.builder()
-                    .statusCode(HttpStatusCodes.STATUS_CODE_BAD_REQUEST)
-                    .status(HttpStatus.BAD_REQUEST)
-                    .message(e.getMessage())
-                    .build();
-        } catch (NullPointerException e) {
-            return Response.builder()
-                    .statusCode(HttpStatusCodes.STATUS_CODE_SERVER_ERROR)
-                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .message(e.getMessage())
-                    .build();
-        }
+    public Response<List<NotificationDTO>> delete(List<Long> notificationsIds) {
+        if (notificationsIds.size() == 0) throw new IllegalArgumentException(ExceptionMessage.EMPTY_LIST.toString());
+        notificationsIds.forEach(notificationId -> {
+            try {
+                Validations.validate(notificationId, Regex.ID.getRegex());
+            } catch (IllegalArgumentException | NullPointerException e) {
+                logger.warn("Couldn't locate this notification");
+            }
+        });
+        List<Notification> remainingNotifications = notificationService.delete(notificationsIds);
+        return Response.<List<NotificationDTO>>builder()
+                .message(SuccessMessage.DELETED.toString())
+                .status(HttpStatus.OK)
+                .statusCode(HttpStatusCodes.STATUS_CODE_NO_CONTENT)
+                .data(NotificationDTO.createNotificationsDTOList(remainingNotifications))
+                .build();
     }
 }
